Added key cache

+++ src/ABC_FileIO.c
@@ -334,7 +334,7 @@ static int readfn(void *handler, char *buf, int size) {
   size_t available = mem->size - mem->pos;
   
   if (size > available) {
-    size = available;
+    size = (int) available;
   }
   memcpy(buf, mem->buffer + mem->pos, sizeof(char) * size);
   mem->pos += size;
@@ -347,7 +347,7 @@ static int writefn(void *handler, const char *buf, int size) {
   size_t available = mem->size - mem->pos;
 
   if (size > available) {
-    size = available;
+    size = (int) available;
   }
   memcpy(mem->buffer + mem->pos, buf, sizeof(char) * size);
   mem->pos += size;
@@ -360,9 +360,9 @@ static fpos_t seekfn(void *handler, fpos_t offset, int whence) {
   fmem_t *mem = handler;
 
   switch (whence) {
-    case SEEK_SET: pos = offset; break;
-    case SEEK_CUR: pos = mem->pos + offset; break;
-    case SEEK_END: pos = mem->size + offset; break;
+    case SEEK_SET: pos = (unsigned long) offset; break;
+    case SEEK_CUR: pos = mem->pos + (unsigned long) offset; break;
+    case SEEK_END: pos = mem->size + (unsigned long) offset; break;
     default: return -1;
   }
